+++ CHANGELOG.md
@@ -4,12 +4,17 @@
 
 ### Breaking Changes
 * Simulation inputs no longer allow setting the id field.
+* Removed nsims_per_server property from Additive class.
 
 ### New Features
 * Default version of Additive Server is 25.1 when creating a client connection.
 * Single bead simulations with thermal history are now supported.
 * Added description field to AdditiveMaterial.
-* Updated the parametric runner to accept a list of simulation ids to be run.
+* Updated the parametric runner to run a list of simulation ids.
+* Asynchronous simulate functionality added.
+* Added apply and list server settings functionality.
+* Parametric study data is updated as simulations are running.
+* Use LOG for logging in Additive object.
 
 ### Bug Fixes
 * Fixed import of microstructure simulations in csv file when random seed is undefined.